Completed Profile Screen
The view of this screen is cool, but I have an issue with navigation. I make tab navigation inside homescreen, so I can't navigate to screens that outside this one. To fix this I have to spend a lot of time which I don't have. So, I will continue work on others tasks(